requirements.txt
a2a-sdk
google-generativeai
uvicorn
python-dotenv
asyncclick
requests
boxsdk[jwt]

gunicorn
google-cloud-aiplatform
google-adk
